Rename elf_loader/ to elf/
